refactor: 디렉토리 구조 일부 변경
- production_schedule에 work_order_assign 추가하여
  생산계획에서 작업지시 분리하고,
  기존 자원관리에 속해있던 '작업자배정'과 '교대유형'을
  생산계획의 work_order_assign/으로 이동
