--- src/main/java/com/megazone/ERPSystem_phase2_Backend/production/repository/production_schedule/work_order/WorkOrderRepository.java
@@ -1,7 +1,0 @@
-package com.megazone.ERPSystem_phase2_Backend.production.repository.production_schedule.work_order;
-
-import com.megazone.ERPSystem_phase2_Backend.production.model.production_schedule.WorkOrder;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WorkOrderRepository extends JpaRepository<WorkOrder, Long>,  WorkOrderRepositoryCustom {
-}
