--- src/main/java/com/megazone/ERPSystem_phase2_Backend/production/repository/production_schedule/work_order/WorkOrderRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.megazone.ERPSystem_phase2_Backend.production.repository.production_schedule.work_order;
-
-public interface WorkOrderRepositoryCustom {
-}
