--- src/main/java/com/megazone/ERPSystem_phase2_Backend/production/repository/production_schedule/work_order/WorkOrderRepositoryImpl.java
@@ -1,9 +1,0 @@
-package com.megazone.ERPSystem_phase2_Backend.production.repository.production_schedule.work_order;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class WorkOrderRepositoryImpl implements WorkOrderRepositoryCustom {
-}
