+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/production/repository/production_schedule/work_order_assign/work_order/WorkOrderRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.megazone.ERPSystem_phase2_Backend.production.repository.production_schedule.work_order_assign.work_order;
+
+public interface WorkOrderRepositoryCustom {
+}
