+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/production/service/production_schedule/work_order_assign/shift_type/ShiftTypeService.java
@@ -1,0 +1,4 @@
+package com.megazone.ERPSystem_phase2_Backend.production.service.production_schedule.work_order_assign.shift_type;
+
+public interface ShiftTypeService {
+}
