+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/production/service/production_schedule/work_order_assign/worker_assignment/WorkerAssignmentService.java
@@ -1,0 +1,4 @@
+package com.megazone.ERPSystem_phase2_Backend.production.service.production_schedule.work_order_assign.worker_assignment;
+
+public interface WorkerAssignmentService {
+}
