--- src/main/java/com/megazone/ERPSystem_phase2_Backend/production/service/resource_data/shift_type/ShiftTypeService.java
@@ -1,4 +1,0 @@
-package com.megazone.ERPSystem_phase2_Backend.production.service.resource_data.shift_type;
-
-public interface ShiftTypeService {
-}
