--- src/main/java/com/megazone/ERPSystem_phase2_Backend/production/service/resource_data/worker_assignment/WorkerAssignmentService.java
@@ -1,4 +1,0 @@
-package com.megazone.ERPSystem_phase2_Backend.production.service.resource_data.worker_assignment;
-
-public interface WorkerAssignmentService {
-}
